fixed issue#63

+++ weixin-mp/src/main/java/com/riversoft/weixin/mp/event/template/JobFinishedEvent.java
@@ -11,17 +11,17 @@ import com.riversoft.weixin.common.event.EventRequest;
 public class JobFinishedEvent extends EventRequest {
 
     @JsonProperty("MsgID")
-    private int msgId;
+    private long msgId;
 
     @JsonProperty("Status")
     @JacksonXmlCData
     private String status;
 
-    public int getMsgId() {
+    public long getMsgId() {
         return msgId;
     }
 
-    public void setMsgId(int msgId) {
+    public void setMsgId(long msgId) {
         this.msgId = msgId;
     }
